Added Blacklisting, logout, minor bug fixes.

+++ resources/item.py
@@ -1,4 +1,4 @@
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity
 from flask_restful import reqparse, Resource
 from models.item import ItemModel
 
@@ -17,7 +17,7 @@ class Item(Resource):
                         required=True,
                         help='Every item needs a store id')
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name=name)
         if item:
@@ -38,7 +38,11 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {"message": "admin privilege is required"}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -59,5 +63,11 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {'items': [item['name'] for item in items],
+                'message': 'More data can be accessed if logged in'}, 200
